+++ _wow_definitions.lua
@@ -4,3 +4,6 @@
 function RunScript(scriptAsText) end
 ---@return number current time
 function GetTime() end
+---@param unit string
+---@return number
+function UnitHealth(unit) end